server/requirements.txt
django
djangorestframework
drf-yasg
django-cors-headers>=3.11.0
djangorestframework-simplejwt